codestyle

+++ scripts/cformat.sh
@@ -12,10 +12,16 @@ clean ()
 
 trap clean EXIT
 
-git ls-files '*.c' | grep -v '^ddtrace/vendor/' | while read filename
+git ls-files '*.c' '*.cpp' '*.h' | grep -v '^ddtrace/vendor/' | while read filename
 do
+
+if [[ "$1" == "update" ]]
+then
+    clang-format --style="{BasedOnStyle: Mozilla, IndentWidth: 4, ColumnLimit: 120}" -i "$filename"
+else
     CFORMAT_TMP=`mktemp`
     clang-format --style="{BasedOnStyle: Mozilla, IndentWidth: 4, ColumnLimit: 120}" "$filename" > "$CFORMAT_TMP"
     diff -u "$filename" "$CFORMAT_TMP"
     rm -f "$CFORMAT_TMP"
+fi
 done
